Rename "templates" directory and subdirectories
Rename the "templates" directory as "example-templates".
It holds example templates which can serve as a starting point
for the user to create their own templates, which are stored in
the "templates" directory. "templates/" will be added to .gitignore.
Change underscores to hyphens in subdirectory names, and change
the settings id "preq" to "deps" as a matter of style.