Set use_mmap option to false for readdlm()
readdlm() uses memory mapping by default. This default currently
prevents files from closing properly when the function finishes.
This lead to errors on NFS, which were fixed in the previous commit.
Setting use_mmap to false reduces the amount of code, and serves as
a reminder of the underlying behavior.