Switch to user named "hdpm" in "deps" and main images
By default the user is root in a docker image. Switching to a non-root
user in the image provides some additional security to the host. The
UID of "hdpm" has been set to 1000. This can be overridden by the "-u"
option of the "docker run" command if needed, such as when mounting
a data volume from a host by a user with a different UID.

The directory "docker/data" in "hdpm" home has been reserved as a docker
volume, and the .gitignore file has been updated accordingly, with a
"work" directory in the "hdpm" folder included for convenience.