Modify style of the path to the amptools package
The path to the amptools package previously ended in "/AmpTools". This
resulted in a problem where the path would automatically be set to just
"AmpTools" instead of "amptools-[VER]/AmpTools" when saving a template.
This commit allows the amptools path to have the same style as the other
package paths.