Add fetch-dist cmd / Docker-image management scripts
A new command called fetch-dist has been added for downloading and
unpacking binary distributions of sim-recon and its dependencies.
The packages are unpacked into the "[top]/.dist" directory, and the
env-setup scripts are updated accordingly.

The new Docker-image management scripts are for automating builds of
Docker images for the supported Linux distributions, and for pushing /
pulling images to/from the Docker hub.

A "cleaning" step has been added to each Dockerfile, which reduces the
uncompressed image size by about 1 GB by deleting src code / Git repos
and build products that are not needed at runtime.