Fix issue with platform-based tarfile selection
The string used to identify the tarfile based on the platform tag needed
some additional characters to make the identification unique.

This commit also adds checks for missing packages, so that an update
will be triggered if any are accidentally deleted. Lastly, it adds a
consistency check between the commit hash obtained from the tarfile name
and the sim-recon success.hdpm file.