Modify default directory structure of pkgs folder
Each package now has a single directory under "pkgs" with subdirectories
for the installed tagged versions and branches. This helps to keep the
pkgs directory better organized. If one wishes to use a different
directory structure it can be modified in settings/paths.txt as usual.

The environment setup script is now named based on the template id,
making it possible to use a previous environment without needing to
regenerate the script.