Provide better integration of binary packages
1. Replace "fetch-dist" with "install" command.
The "install" command fetches the binary distribution as before, but
after that it links the directories housing binaries in the "pkgs/.dist"
directory to visible directories in the "pkgs" directory. It will only
install a package if the requested package version is not already
installed in the "pkgs" directory.

2. Remove "home-dev"/"dist" build templates.
These templates are no longer necessary. The "master" template in
combination with the "install" command is sufficient.