Avoid incompatibility between cling and devtoolset
Building ROOT 6 (cling) with gcc from devtoolset is not officially
supported because of an incompatibility
(https://sft.its.cern.ch/jira/browse/ROOT-6207).
The recommended build method, with cmake, produces defective products.
If you try to join strings (with cling), for example, there is a
missing symbol error.

A workaround is to use the old build method instead:
./configure; make
Doing so results in cling using the builtin (compatible) string_view.